requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery[redis]==5.3.6
certifi==2023.11.17
channels-redis==4.2.0
channels[daphne]==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
kombu==5.3.5
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
phonenumbers==8.13.27
